requirements.txt
discord.py==2.3.2
python_version >= "3.10"
google-generativeai==0.5.2
